requirements.txt
flask[async]
pyppeteer
python-dotenv